[expo-image][android] Implement our own ExpoImageView
Then we'll be able to manage some lifecycle events more easily in a separate class.
